refactor: restructure rai_bench
