+++ src/rai_bench/README.md
@@ -1,20 +1,10 @@
-## RAI Benchmark
-
-### Description
+# RAI Benchmarks
 
 The RAI Bench is a package including benchmarks and providing frame for creating new benchmarks
 
-### Frame Components
+## Manipulation O3DE Benchmark
 
-- `Task`
-- `Scenario`
-- `Benchmark`
-
-For more information about these classes go to -> [benchmark_model](./rai_bench/benchmark_model.py)
-
-### O3DE Test Benchmark
-
-The O3DE Test Benchmark [o3de_test_benchmark_module](./rai_bench/o3de_test_bench/) provides tasks and scene configurations for robotic arm manipulation task. The tasks use a common `ManipulationTask` logic and can be parameterized, which allows for many task variants. The current tasks include:
+The Manipulation O3DE Benchmark [manipulation_o3de_benchmark_module](./rai_bench/manipulation_o3de_bench) provides tasks and scene configurations for robotic arm manipulation simulation in O3DE. The tasks use a common `ManipulationTask` logic and can be parameterized, which allows for many task variants. The current tasks include:
 
 - **MoveObjectToLeftTask**
 - **GroupObjectsTask**
@@ -26,6 +16,14 @@ The result of a task is a value between 0 and 1, calculated like initially_mispl
 
 Current O3DE simulation binaries:
 
+### Frame Components
+
+- `Task`
+- `Scenario`
+- `Benchmark`
+
+For more information about these classes go to -> [benchmark](./rai_bench/manipulation_o3de_bench/benchmark.py)
+
 ### Running
 
 1. Download O3DE simulation binary and unzip it.
@@ -35,13 +33,13 @@ Current O3DE simulation binaries:
 
 2. Follow step 2 from [Manipulation demo Setup section](../../docs/demos/manipulation.md#setup)
 
-3. Adjust the path to the binary in: [o3de_config.yaml](./rai_bench/o3de_test_bench/configs/o3de_config.yaml)
+3. Adjust the path to the binary in: [o3de_config.yaml](./rai_bench/examples/manipulation_o3de/configs/o3de_config.yaml)
 4. Run benchmark with:
 
    ```bash
    cd rai
    source setup_shell.sh
-   python src/rai_bench/rai_bench/examples/o3de_test_benchmark.py
+   python src/rai_bench/rai_bench/examples/manipulation_o3de/main.py
    ```
 
 > [!NOTE]
@@ -50,7 +48,7 @@ Current O3DE simulation binaries:
 
 ### Example usage
 
-Example of how to load scenes, define scenarios and run benchmark can be found in [o3de_test_benchmark_example](./rai_bench/examples/o3de_test_benchmark.py)
+Example of how to load scenes, define scenarios and run benchmark can be found in [manipulation_o3de_benchmark_example](rai_bench/examples/manipulation_o3de/main.py)
 
 Scenarios can be loaded manually like:
 
@@ -73,7 +71,7 @@ scenarios = Benchmark.create_scenarios(
 
 which will result in list of scenarios with combination of every possible task and scene(task decides if scene config is suitable for it).
 
-or can be imported from exisitng packets [scenarios_packets](./rai_bench/o3de_test_bench/scenarios.py):
+or can be imported from exisitng packets [scenarios_packets](rai_bench/examples/manipulation_o3de/scenarios.py):
 
 ```python
 t_scenarios = trivial_scenarios(
@@ -94,28 +92,28 @@ vh_scenarios = very_hard_scenarios(
 ```
 
 which are grouped by their subjective difficulty. For now there are 10 trivial, 42 easy, 23 medium, 38 hard and 47 very hard scenarios.
-Check docstrings and code in [scenarios_packets](./rai_bench/o3de_test_bench/scenarios.py) if you want to know how scenarios are assigned to difficulty level.
+Check docstrings and code in [scenarios_packets](rai_bench/examples/manipulation_o3de/scenarios.py) if you want to know how scenarios are assigned to difficulty level.
 
 ### Development
 
-When creating new task or changing existing ones, make sure to add unit tests for score calculation in [rai_bench_tests](../../tests/rai_bench/).
+When creating new task or changing existing ones, make sure to add unit tests for score calculation in [rai_bench_tests](../../tests/rai_bench/manipulation_o3de/tasks/).
 This applies also when you are adding or changing the helper methods in `Task` or `ManipulationTask`.
 
 The number of scenarios can be easily extened without writing new tasks, by increasing number of variants of the same task and adding more simulation configs but it won't improve variety of scenarios as much as creating new tasks.
 
-### Tool Calling Agent Benchmark
+## Tool Calling Agent Benchmark
 
 The Tool Calling Agent Benchmark is the benchmark for LangChain tool calling agents. It includes a set of tasks and a benchmark that evaluates the performance of the agent on those tasks by verifying the correctness of the tool calls requested by the agent. The benchmark is integrated with LangSmith and Langfuse tracing backends to easily track the performance of the agents.
 
 #### Frame Components
 
-- [Tool Calling Agent Benchmark](rai_bench/tool_calling_agent_bench/agent_bench.py) - Benchmark for LangChain tool calling agents
-- [Tasks Interfaces](rai_bench/tool_calling_agent_bench/agent_tasks_interfaces.py) - Interfaces for tool calling agent tasks
+- [Tool Calling Agent Benchmark](rai_bench/tool_calling_agent_bench/benchmark.py) - Benchmark for LangChain tool calling agents
+- [Tasks Interfaces](rai_bench/tool_calling_agent_bench/interfaces.py) - Interfaces for tool calling agent tasks
 - [Scores tracing](rai_bench/tool_calling_agent_bench/scores_tracing.py) - Component handling sending scores to tracing backends
 
 #### Benchmark Example with ROS2 Tools
 
-[tool_calling_agent_test_bench.py](rai_bench/examples/tool_calling_agent_test_bench.py) - Script providing benchmark on tasks based on the ROS2 tools usage.
+[tool_calling_agent_test_bench.py](rai_bench/examples/tool_calling_agent/main.py) - Script providing benchmark on tasks based on the ROS2 tools usage.
 
 To set up tracing backends, please follow the instructions in the [tracing.md](../../docs/tracing.md) document.
 
@@ -124,7 +122,7 @@ To run the benchmark:
 ```bash
 cd rai
 source setup_shell.sh
-python src/rai_bench/rai_bench/examples/tool_calling_agent_test_bench.py
+python src/rai_bench/rai_bench/examples/tool_calling_agent/main.py
 
 > [!NOTE]
 > The `simple_model` from [config.toml](../../config.toml) is currently set up in the example benchmark script. Change it to `complex_model` in the script if needed.
