+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/1bc_1rc_1yc.yaml
