+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/1carrot_1a_2t_1bc_1rc_3yc_stacked.yaml
