+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/1carrot_1t_1rc.yaml
