+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/1rc_2bc_3yc.yaml
