+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/1t.yaml
