+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/2a_1bc.yaml
