+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/2carrots_1a_1t_1bc_1rc_1yc_1corn.yaml
