+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/2rc_3bc_4yc_stacked.yaml
