+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/2t_3a_1corn_2rc.yaml
