+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/2yc_1bc_1rc.yaml
