+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/3a_4t_2bc.yaml
