+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/3carrots_1a_1t_2bc_2yc.yaml
