+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/3carrots_1a_2bc_1rc_1yc_1corn.yaml
