+++ src/rai_bench/rai_bench/examples/manipulation_o3de/configs/3rc_3bc_stacked.yaml
