+++ src/rai_bench/rai_bench/examples/tool_calling_agent/tasks.py
@@ -14,7 +14,7 @@
 
 from typing import Sequence
 
-from rai_bench.tool_calling_agent_bench.agent_tasks_interfaces import (
+from rai_bench.tool_calling_agent_bench.interfaces import (
     ToolCallingAgentTask,
 )
 from rai_bench.tool_calling_agent_bench.ros2_agent_tasks import (
