--- src/rai_bench/rai_bench/o3de_test_bench/configs/o3de_config.yaml
@@ -1,19 +1,0 @@
-binary_path: /path/to/RAIManipulationDemo/RAIManipulationDemo.GameLauncher
-level: RoboticManipulationBenchmark
-robotic_stack_command: ros2 launch examples/manipulation-demo-no-binary.launch.py
-required_simulation_ros2_interfaces:
-  services:
-    - /spawn_entity
-    - /delete_entity
-  topics:
-    - /color_image5
-    - /depth_image5
-    - /color_camera_info5
-  actions: []
-required_robotic_ros2_interfaces:
-  services:
-    - /grounding_dino_classify
-    - /grounded_sam_segment
-    - /manipulator_move_to
-  topics: []
-  actions: []
