+++ src/rai_bench/rai_bench/tool_calling_agent_bench/ros2_agent_tasks.py
@@ -23,7 +23,7 @@ from langchain_core.messages.tool import ToolCall
 from langchain_core.tools import BaseTool
 from rai.tools.ros2 import MoveToPointToolInput
 
-from rai_bench.tool_calling_agent_bench.agent_tasks_interfaces import (
+from rai_bench.tool_calling_agent_bench.interfaces import (
     ROS2ToolCallingAgentTask,
 )
 from rai_bench.tool_calling_agent_bench.mocked_tools import (
