+++ src/rai_core/rai/initialization/model_initialization.py
@@ -15,12 +15,15 @@
 import logging
 import os
 from dataclasses import dataclass
-from typing import List, Literal, Optional, Tuple, cast
+from typing import Any, List, Literal, Optional, Tuple, cast
 
 import coloredlogs
 import tomli
+from langchain_aws import ChatBedrock
 from langchain_core.callbacks.base import BaseCallbackHandler
 from langchain_core.tracers.langchain import LangChainTracer
+from langchain_ollama import ChatOllama
+from langchain_openai import ChatOpenAI
 from langsmith import Client
 
 logger = logging.getLogger(__name__)
