+++ tests/rai_bench/manipulation_o3de/tasks/test_group_objects_task.py
@@ -14,7 +14,7 @@
 
 from typing import List
 
-from rai_bench.o3de_test_bench.tasks import GroupObjectsTask
+from rai_bench.manipulation_o3de_bench.tasks import GroupObjectsTask
 from rai_sim.simulation_bridge import Entity
 from tests.rai_bench.conftest import create_entity
