+++ tests/rai_bench/manipulation_o3de/tasks/test_task.py
